+++ ta4j-core/src/test/java/org/ta4j/core/indicators/KSTIndicatorTest.java
@@ -23,16 +23,16 @@
  */
 package org.ta4j.core.indicators;
 
+import static org.ta4j.core.TestUtils.assertNumEquals;
+
+import java.util.function.Function;
+
 import org.junit.Before;
 import org.junit.Test;
 import org.ta4j.core.Indicator;
 import org.ta4j.core.indicators.helpers.ClosePriceIndicator;
 import org.ta4j.core.mocks.MockBarSeries;
 import org.ta4j.core.num.Num;
-
-import java.util.function.Function;
-
-import static org.ta4j.core.TestUtils.assertNumEquals;
 
 public class KSTIndicatorTest extends AbstractIndicatorTest<Indicator<Num>, Num> {
     private MockBarSeries data;