+++ ta4j-core/src/main/java/org/ta4j/core/num/DoubleNum.java
@@ -35,8 +35,6 @@ import java.util.function.Function;
  */
 public class DoubleNum implements Num {
 
-    private static final long serialVersionUID = -2611177221813615070L;
-
     private static final DoubleNum ZERO = DoubleNum.valueOf(0);
     private static final DoubleNum ONE = DoubleNum.valueOf(1);
     private static final DoubleNum HUNDRED = DoubleNum.valueOf(100);