+++ CHANGELOG.md
@@ -7,9 +7,9 @@ Changelog for `ta4j`, roughly following [keepachangelog.com](http://keepachangel
 ### Fixed
 
 ### Changed
-- **Num** removed Serializable
 
 ### Removed/Deprecated
+- **Num** removed Serializable
 
 ### Added